requirements.txt
cuscustomtkinter
sqlite
datetime
tkinter